% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xml.R
\name{build_xml}
\alias{build_xml}
\title{Construct XML schema to commuticate to RSAT.}
\usage{
build_xml(method, parameters = NULL)
}
\arguments{
\item{method}{Name of the method to be used from RSAT}

\item{parameters}{List of parameters provided to method}
}
\value{
An XML schema
}
\description{
This function creates an XML schema to be used by `POST` function.
}
\examples{
request <- build_xml("supported_organisms",
                     parameters = list(return = "name,source",
                                       source = "ensembl")
                   )
}
\author{
José Alquicira Hernández & Joselyn Chávez
}
