% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCondition.R
\name{BuildCondition}
\alias{BuildCondition}
\title{Constructs logical condition to query database}
\usage{
BuildCondition(filters, dataset, operator)
}
\arguments{
\item{filters}{A list with values to restrict the query. The names of all elements in this list correspond to the attributes to bre retrieved.}

\item{dataset}{A string containing the name of the dataset of interest}

\item{operator}{A string indicading if all conditions -\code{filters}- (AND) or some of them (OR) should be met}
}
\value{
A string with a sql logical condition to query the dataset
}
\description{
Given a list of filters, this function builds a logical condition to query database
}
\examples{
BuildCondition(filters = list(network_type = "TF-GENE",
                         regulator_name = "AraC"),
                          dataset = "NETWORK",
                          operator = "AND")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
