% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildXml.R
\name{BuildXml}
\alias{BuildXml}
\title{Construct XML schema to commuticate to RSAT.}
\usage{
BuildXml(method, parameters = NULL)
}
\arguments{
\item{method}{Name of the method to be used from RSAT}

\item{parameters}{List of parameters provided to method}
}
\value{
an XML schema
}
\description{
This function creates an XML schema to be used by `POST` function.
}
\examples{
request <- BuildXml("supported_organisms",
                     parameters = list(return = "name,source",
                     source = "ensembl"))
}
\author{
José Alquicira Hernández
}
