% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{convert_to_biostrings}
\alias{convert_to_biostrings}
\title{Function to convert output of regulondb queries to Biostrings objects}
\usage{
convert_to_biostrings(regulondb_result, seq_type = "DNA")
}
\arguments{
\item{regulondb_result}{A regulon_result object.}

\item{seq_type}{A character string with either DNA or protein, specyfing what}
}
\value{A Biostrings object}
\description{
This function converts, when possible, a regulon_result object into a Biostrings object.
}
\author{
Alejandro Reyes
}
