% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{convert_to_granges}
\alias{convert_to_granges}
\title{Function to convert output of regulondb queries to Bioconductor S4 objects}
\usage{
convert_to_granges(regulondb_result)
}
\arguments{
\item{regulondb_result}{A data frame that resulted from a call to the function get_dataset.}
}
\description{
This function
}
\author{
Alejandro Reyes
}
\keyword{attributes,}
\keyword{data}
\keyword{filters,}
\keyword{retrieval,}
