% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_pattern.R
\name{dna_pattern}
\alias{dna_pattern}
\title{Searches all occurrences of a pattern within DNA sequences via RSAT.}
\usage{
dna_pattern(
  sequence,
  format = "fasta",
  substitution = NULL,
  pattern,
  id = NULL,
  origin = NULL,
  noov = NULL,
  both.strand = TRUE,
  sort.oligo = NULL,
  threshold = NULL
)
}
\arguments{
\item{sequence}{Input sequence}

\item{format}{Input sequence format. Supported: 'IG' (Intelligenetics), 'WC' (wconsensus), 'raw', 'fasta'. Defaul is 'fasta'.}

\item{substitution}{Number of substitutions allowed.}

\item{pattern}{Pattern to match.}

\item{id}{Pattern identifier.}

\item{origin}{Origin for the calculation of positions (0 for end of sequence).}

\item{noov}{No overlapping of oligos allowed if TRUE. Disable the detection of overlapping matches for self-overlapping patterns (e.g. TATATA, GATAGA).}

\item{both.strand}{Oligonucleotide occurrences found on both stands are summed (TRUE) or not (FALSE). Default is TRUE}

\item{sort.oligo}{sort oligomers according to overrepresentation if TRUE.}

\item{threshold}{Threshold on match count.}
}
\value{
A data frame with columns
\itemize{
\item {patternID}
\item{strand}
\item{pattern}
\item{seqID}
\item{start}
\item{end}
\item{matching_seq}
\item{score}
}
}
\description{
Returns position of pattern match in the input sequence .
}
\examples{
# Find pattern in both strands of input sequence
res <- dna_pattern(sequence = "TCGCAATCTACTTACTATTGCTTACTATTGCTTACTATTGCGAGCGCAGA",
                  format = "raw",
                  pattern = "TCGCAAT")
res <- dna_pattern(sequence = ">seq1\nTCGCAATCTACTTACTATTGCTTACTATTGCTTACTATTGCGAGCGCAGA",
                  pattern = "TCGCAAT",
                  both.strand = F,
                  id = "patt_1")
cat(res)
}
\author{
Joselyn Chávez
}
