% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_database.R
\name{download_database}
\alias{download_database}
\title{Download regulondb database from online file}
\usage{
download_database()
}
\value{
local file regulondb_sqlite3.db
}
\description{
Download regulondb database from online file
}
\examples{
download_database()
}
