% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExistingPartialMatch.R
\name{ExistingPartialMatch}
\alias{ExistingPartialMatch}
\title{Constructs a particular logical condition to query database to add google search}
\usage{
ExistingPartialMatch(filters, partialmatch, operator)
}
\arguments{
\item{filters}{A list with values to restrict the query. The partialmatchs of all elements in this list correspond to the attributes to bre retrieved.}

\item{operator}{A string indicading if all conditions -\code{filters}- (AND) or some of them (OR) should be met}

\item{interv}{the filters that are going to be in a intervale}
}
\value{
A string with a sql logical condition to query the dataset
}
\description{
Given a list of filters, this function builds a logical condition to query database using intervals
}
\examples{
BuildCondition(filters = list(network_type = "TF-GENE",
                         regulator_partialmatch = "AraC"),
                          dataset = "NETWORK",
                          operator = "AND")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
