% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graphic_points.R
\name{generate_graphic_points}
\alias{generate_graphic_points}
\title{Generation of a scatter plot}
\usage{
generate_graphic_points(
  df,
  column_x,
  column_y,
  x_axis_title,
  y_axis_title,
  plot_title
)
}
\arguments{
\item{df}{The dataframe containing the data for the plot.}

\item{column_x}{The name of the column to be used on the x-axis.}

\item{column_y}{The name of the column to be used on the y-axis.}

\item{x_axis_title}{The title for the x-axis.}

\item{y_axis_title}{The title for the y-axis.}

\item{plot_title}{The title for the plot.}
}
\value{
A scatter plot generated using ggplot2.
}
\description{
This function uses the ggplot2 library to generate
a scatter plot based on data from a dataframe.
It allows specifying the columns for the x and y axes,
as well as the titles for the axes and the plot.
}
\examples{
Example 1: Generate a scatter plot of the number of genes by operons
generate_graphic_points_genes_by_operon <- generate_graphic_points(get_general_frequency_genes,cantidad,n_operon,"genes by operon","operon frequency","Genes by operon")

Example 2: Generate a scatter plot of the number of genes by tfs
generate_graphic_points_genes_by_tfs <- generate_graphic_points(get_general_frequency_genes_by_tf,	n_tf, cantidad, "Number regulated genes ", "Number of tf","Number of genes regulated by tf")



}
