% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequency.R
\name{get_frequency}
\alias{get_frequency}
\title{Obtains the frequency of repeated values in a column.}
\usage{
get_frequency(df, column_name, column_total)
}
\arguments{
\item{df}{The dataset being worked with.}

\item{column_name}{Column of the dataframe by which the data grouping will be created
to get the frequency}

\item{column_total}{Name of the column where the repeated data will be summed.}
}
\value{
Dataset with the frequency of the data from the second parameter (column_name).
}
\description{
will get the frequency with which each element of the column is repeated
given as the second parameter of the function.
}
\examples{
# Example 1: Obtain the frequency of the number of transcriptionUnits by operon
frecuency_tus <- make_frecuency(df = tus_by_operon, column_name = cantidad, column_total = n_operon)
# Example 2:
frequency_by_genes <- make_frequency(total_genes_by_tf, cantidad, n_tf)
}
\author{
Pina Acosta Daniel, Valle Mondragon Sergio.
}
