% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Return summary of gene regulation.}
\usage{
get_summary(geneRegulation)
}
\arguments{
\item{geneRegulation}{Result from \code{\link{GetGeneRegulation}} or vector of genes}
}
\value{
A data frame with the following columns:
\itemize{
\item The name or gene of TF
\item Regulated Genes per TF
\item Percent of regulated genes per TF
\item positive, negative or dual regulation
\item Name(s) of regulated genes
}
}
\description{
This function takes the output of \code{\link{GetGeneRegulation}} with format multirow,
onerow or table, or a vector with genes and retrieves information about the TFs and their regulated genes
}
\examples{
# Retrieve summary of  araC's regulation

araC_regulation<- get_gene_regulators(genes = c("araC"),
                                    format = "multirow",
                                    output.type = "TF")
get_summary(araC_regulation)

# Retrieve summary of genes 'araC' and 'modB'

get_summary(geneRegulation = c("araC", "modB"))
get_summary(geneRegulation = c("ECK120000050", "modB"))
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chávez
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{summary,}
