% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttr.R
\name{getAttr}
\alias{getAttr}
\title{Get Attributes}
\usage{
getAttr(attributes = NULL, filters = NULL, values = NULL,
  dataset = NULL, operator = "AND")
}
\arguments{
\item{attributes}{Attributes to be retrieved.  Defaults to NULL.
filters   Filters to be used.  Defaults to NULL.
values Values to filter with.  Defaults to NULL.
dataset Dataset to retrieve from. Defaults to NULL.
operator  Logical operator for the filters: AND / OR.  Defaults to AND.}
}
\description{
Retrieve attributes from datasets.
}
\examples{
getAttr(attributes=c("gene_id"), dataset="GENE")
}
\keyword{attributes,}
\keyword{data}
\keyword{retrieval,}

