% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBindingSites.R
\name{GetBindingSites}
\alias{GetBindingSites}
\title{Get the binding sites for a Transcription Factor (TF)}
\usage{
GetBindingSites(TF, seq.format = "table", evidence = NULL)
}
\arguments{
\item{TF}{name of the transcription factor of interest}

\item{seq.format}{Default: \code{table}. Supported: \code{fasta}, \code{wconsensus}.}
}
\value{
A dataframe with the following columns corresponding to the TFBSs associated with the TF.
If seq.format = "fasta", returns a string with the TFBS sequences with '\n' breakline character.
If the transcription factor does not exist, returns NA.
\itemize{
\item ID (identifier)
\item left position
\item right position
\item strand
\item sequence
}
}
\description{
Retrieve the binding sites and genome location for a given transcription factor.
}
\examples{
# Get the binding sites for AraC

GetBindingSites(TF = "AraC")

# Get the fasta binding sites sequences for AraC

GetBindingSites(TF = "AraC", seq.format = "fasta")
}
\author{
José Alquicira Hernández, Jacques van Helden, Joselyn Chávez
}
