% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatabase.R
\name{GetDatabase}
\alias{GetDatabase}
\title{Download up to date RegulonDB lite from GitHub}
\usage{
GetDatabase()
}
\description{
This function connects to GitHub to download the most up to date version of the Regulon Database. No parameters are provided. The database is stored in the inst/extdata directory of the package.
}
\examples{
GetDatabase()
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, Josá Alquicira Hernández
}
\keyword{data}
\keyword{database}
\keyword{datasets,}
\keyword{retrieval,}
