% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneRegulation.R
\name{GetGeneRegulation}
\alias{GetGeneRegulation}
\title{Get regulation}
\usage{
GetGeneRegulation(genes, format = "multirow", output.type = "TF")
}
\arguments{
\item{genes}{List of genes or GIs.}

\item{format}{Output format of data:}

\item{output.type}{Should regulators be represented as TFs of genes?}
}
\description{
Given a list of genes (names, bnumbers or GIs), get all transcription factors that regulate them.
}
\examples{
# Retrieve regulation of araC

GetGeneRegulation(genes = c("araC"), 
format = "multirow", 
output.type = "TF")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{tf,}
