% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetwork.R
\name{GetNetwork}
\alias{GetNetwork}
\title{GetNetwork}
\usage{
GetNetwork(type = "TF-GENE")
}
\arguments{
\item{type}{TF-GENE, TF-TF, GENE-GENE}
}
\description{
Return complete regulatory network.
}
\examples{
# Retrieve regulation of araC

GetNetwork(type="TF-GENE")



}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{tf,}
