% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSummary.R
\name{GetSummary}
\alias{GetSummary}
\title{GetSummary}
\usage{
GetSummary(regulation)
}
\arguments{
\item{regulation}{The input of this function is de output of GetGeneRegulation()}
}
\description{
Return summary of regulation.
}
\examples{
# Retrieve summary of  araC

araC_regulation<- GetGeneRegulation(genes = c("araC"),
format = "multirow",
output.type = "TF")

GetSummary(regulation=araC_regulation)
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{summary,}
