% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_attributes.R
\name{list_attributes}
\alias{list_attributes}
\title{List attributes from a dataset}
\usage{
list_attributes(dataset, comments = FALSE)
}
\arguments{
\item{dataset}{Dataset of interest}

\item{coments}{Default FALSE. If TRUE, displays the description of each attribute}
}
\value{
A data frame with two columns:
\itemize{
\item \code{attribute}. Name of the attribute
\item \code{description}. Description of attribute
}
}
\description{
List all attributes and their description of a dataset from RegulonDB. The result of this function may
be used as parameter 'values' in \code{\link{GetAttr}} function.
}
\examples{
list_attributes("TF")
list_attributes("OPERON", comments = TRUE)
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chavez
}
\keyword{attributes,}
\keyword{data}
\keyword{retrieval,}
