% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets in RegulonDB database}
\usage{
list_datasets()
}
\description{
This function returns a vector of all available datasets. No parameters are provided.
}
\examples{
list_datasets()
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, Josá Alquicira Hernández
}
\keyword{data}
\keyword{database,}
\keyword{datasets,}
\keyword{retrieval,}
