% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAttributes.R
\name{listAttributes}
\alias{listAttributes}
\title{List Attributes}
\usage{
listAttributes(dataset)
}
\arguments{
\item{dataset}{Dataset to retrieve from. Defaults to NULL.}
}
\description{
List attributes from datasets.
}
\examples{
listAttributes("GENE")
}
\keyword{attributes,}
\keyword{data}
\keyword{retrieval,}

