% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListDatasets.R
\name{ListDatasets}
\alias{ListDatasets}
\title{List available datasets in RegulonDB database}
\usage{
ListDatasets()
}
\description{
This function returns a vector of all available datasets. No parameters are provided.
}
\examples{
listDatasets()
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, Josá Alquicira Hernández
}
\keyword{data}
\keyword{database,}
\keyword{datasets,}
\keyword{retrieval,}
