% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_granges.R
\name{plot_GRanges}
\alias{plot_GRanges}
\title{Plot elements within a genome region}
\usage{
plot_GRanges(regulondb, from, to, plot_map = TRUE)
}
\arguments{
\item{regulondb}{A regulondb object.}

\item{from}{left position.}

\item{to}{right position}

\item{plot_map}{By default TRUE, if FALSE returns a GRanges object with the elements found.}
}
\value{
A plot with genomic elements found within a genome region, including genes and regulators.
}
\description{
Plot elements within a genome region
}
\examples{
## Download the database if necessary
if(!file.exists(file.path(tempdir(), 'regulondb_sqlite3.db'))) {
    download_database(tempdir())
}

## Build the regulondb object
e_coli_regulondb <-
    regulondb(
        database_path = file.path(tempdir(), "regulondb_sqlite3.db"),
        organism = "chr",
        database_version = "1",
        genome_version = "1"
    )
plot_GRanges(e_coli_regulondb, from = 5000, to = 10000)
}
\author{
Joselyn Chavez
}
