% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulondb-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,regulondb-method}
\title{Methods for regulondb objects}
\usage{
\S4method{show}{regulondb}(object)
}
\arguments{
\item{object}{A regulondb object}
}
\value{
A \link[=regulondb]{regulondb} object.
}
\description{
Methods for regulondb objects
}
