% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulondb-classes.R
\name{regulondb}
\alias{regulondb}
\alias{build_regulondb}
\title{Constructor function of a regulondb class}
\usage{
regulondb(database_path, organism, genome_version, database_version)
}
\description{
The build_regulondb function is a constructor function of a
regulondb
class.
}
\section{Slots}{

\describe{
\item{\code{organism}}{A character vector with the name of the organism of the
database.}

\item{\code{genome_version}}{A character vector with the version of the genome build.}

\item{\code{database_version}}{A character vector with the version of regulondb
build.}
}}

\examples{

## Download the database if necessary
if(!file.exists(file.path(tempdir(), 'regulondb-sqlite3.db'))) {
    download_database(tempdir())
}

## Build a regulondb object
e_coli_regulondb <-
    regulondb(
        database_path = file.path(tempdir(), 'regulondb-sqlite3.db'),
        organism = "E.coli",
        database_version = "1",
        genome_version = "1"
    )

}
