% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regutools.R
\docType{package}
\name{regutools}
\alias{regutools}
\title{regutools package}
\description{
regutools extracts and processes data from RegulonDB
}
\section{Maintainer}{
 Joselyn Chavez
}

\section{See Also}{

Report bugs at https://github.com/ComunidadBioInfo/regutools.git/issues
}

\author{
Joselyn Chavez, Jose Alquicira Hernandez, Carmina Barberena Jonas, Jesus Emiliano Sotelo Fonseca
}
