% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAT.R
\name{RSAT}
\alias{RSAT}
\title{Call RSAT.}
\usage{
RSAT(method, parameters = NULL)
}
\arguments{
\item{method}{Name of the method to be used from RSAT}

\item{parameters}{List of parameters provided to method}
}
\value{
an R object with results retrieved from RSAT
}
\description{
This function access RSAT via SOAP.
}
\examples{

# Get 50 random protein sequences of length 100 in fasta format. Use number 300 as seed.

RSAT(method = "random_seq",
    parameters = list(sequence_length = 100,
                      repetition = 50,
                      type = "protein",
                      seed = 300,
                      format = "fasta"))
}
\author{
José Alquicira Hernández
}
