% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_database.R
\name{connect_database}
\alias{connect_database}
\title{Connect to the regulondb database}
\usage{
connect_database(
  path = tempdir(),
  overwrite = FALSE,
  ah = AnnotationHub::AnnotationHub()
)
}
\arguments{
\item{path}{The directory where the database will be downloaded to.}

\item{overwrite}{\code{logical(1)} specifying whether to overwrite the database
file.}

\item{ah}{An \code{AnnotationtHub} object
\link[AnnotationHub:AnnotationHub-class]{AnnotationHub-class}. Can be \code{NULL}
if you want to force to use the backup download mechanism.}
}
\value{
An \link[RSQLite:SQLiteConnection-class]{SQLiteConnection-class}
connection to the RegulonDB database.
}
\description{
If the file does not exist, this function downloads
the RegulonDB SQLite database file into the \code{path} prior to loading it.
}
\examples{

## Connect to the RegulonDB database if necessary
if(!exists('regulondb_conn')) regulondb_conn <- connect_database()

## Connect to the database without using AnnotationHub
regulondb_conn_noAH <- connect_database(ah = NULL, overwrite = TRUE)

}
