% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{convert_to_granges}
\alias{convert_to_granges}
\title{Function to convert output of regulondb queries to GenomicRanges objects}
\usage{
convert_to_granges(regulondb_result)
}
\arguments{
\item{regulondb_result}{A regulon_result object.}
}
\description{
This function converts, when possible, a regulon_result object into a GRanges object.
}
\author{
Alejandro Reyes
}
