% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_database.R
\name{download_database}
\alias{download_database}
\title{Download regulondb database}
\usage{
download_database(path = ".", overwrite = FALSE)
}
\value{
local file regulondb_sqlite3.db
}
\description{
If the file does not exist, this function downloads regulondb_sqlite3.db into the extdata directory of the regutools package.
}
\examples{
download_database()
}
