% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_database.R
\name{download_database}
\alias{download_database}
\title{Download regulondb database}
\usage{
download_database(path = tempdir(), overwrite = FALSE)
}
\arguments{
\item{path}{The directory where the database will be downloaded to.}

\item{overwrite}{\code{logical(1)} specifying whether to overwrite the database
file.}
}
\value{
local file regulondb_sqlite3.db
}
\description{
If the file does not exist, this function downloads
regulondb_sqlite3.db into the extdata directory of the regutools package.
}
\examples{

## Download the database if necessary
if(!file.exists(file.path(tempdir(), 'regulondb_sqlite3.db'))) {
    download_database(tempdir())
}

}
