% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExistingPartialMatch.R
\name{ExistingPartialMatch}
\alias{ExistingPartialMatch}
\title{Constructs a logical condition to query database}
\usage{
ExistingPartialMatch(filters, partialmatch, operator)
}
\arguments{
\item{filters}{List of filters to be used. The names should correspond to the attribute and the values correspond to the condition for selection.}

\item{partialmatch}{name of the condition(s) with a string pattern for full or partial match in the query}

\item{operator}{A string indicating if all the filters (AND) or some of them (OR) should be met}
}
\value{
A sql logical condition to query the dataset
}
\description{
Given a list of filters, this function builds a logical condition to query database using intervals.
The output is used in \code{\link{ExistingIntervals}} and \code{\link{NonExistingIntervals}}.
}
\examples{
ExistingPartialMatch(filters = list(name=c("ara"),
                                    strand=c("forward"),
                                    posright=c("2000","40000")),
                     partialmatch = "name",
                     operator = "AND")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
