% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binding_sites.R
\name{get_binding_sites}
\alias{get_binding_sites}
\title{Get the binding sites for a Transcription Factor (TF)}
\usage{
get_binding_sites(regulondb, TF, seq.format = "table")
}
\arguments{
\item{regulondb}{A regulondb object.}

\item{TF}{name of the transcription factor.}

\item{seq.format}{Default: \code{table}. Supported: \code{fasta}, \code{wconsensus}.}
}
\value{
A dataframe with the following columns corresponding to the TFBSs
associated with the TF.
If seq.format = "fasta", returns a string with the TFBS sequences with
breakline character.
If the transcription factor does not exist, returns \code{NA}.
\itemize{
\item ID (identifier)
\item left position
\item right position
\item strand
\item sequence
}
}
\description{
Retrieve the binding sites and genome location for a given
transcription factor.
}
\examples{
## Download the database if necessary
if(!file.exists(file.path(tempdir(), 'regulondb-sqlite3.db'))) {
    download_database(tempdir())
}

## Build the regulon db object
e_coli_regulondb <-
    regulondb(
        database_path = file.path(tempdir(), "regulondb_sqlite3.db"),
        organism = "E.coli",
        database_version = "1",
        genome_version = "1"
    )

## Get the binding sites for AraC
get_binding_sites(e_coli_regulondb, TF = "AraC")

## Get the fasta binding sites sequences for AraC
fasta.seq <- get_binding_sites(e_coli_regulondb,
    TF = "AraC", seq.format = "fasta")
cat(fasta.seq)

}
\author{
José Alquicira Hernández, Jacques van Helden, Joselyn Chávez
}
