% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttr.R
\name{GetAttr}
\alias{GetAttr}
\title{Extract data from RegulonDB}
\usage{
GetAttr(dataset = NULL, attributes = NULL, filters = NULL,
  and = TRUE, interval = NULL, partialmatch = NULL)
}
\arguments{
\item{dataset}{Dataset of interest.}

\item{attributes}{Vector of attributes to be retrieved}

\item{filters}{List of filters to be used. The names should correspond to the attribute and the values correspond to the condition for selection.}

\item{and}{Logical argument. If FALSE, filters will be considered under the "OR" operator}

\item{interval}{the filters whose values will be considered as interval}

\item{partialmatch}{name of the condition(s) with a string pattern for full or partial match in the query}
}
\description{
This function retrieves data from RegulonDB. Attributes from datasets can be selected and filtered.
}
\examples{
# Obtain all the information from the "GENE" dataset
GetAttr(dataset="GENE")

# Get the attributes posright and name from the "GENE" dataset
GetAttr(dataset="GENE", attributes=c("posright","name") )

# From "GENE" dataset, get the gene name, strand, posright, product name and id of all genes regulated
  with name like "ara", strand as "forward" with a position rigth between 2000 and 40000

GetAttr(dataset = "GENE",
        attributes = c("name", "strand", "posright", "product_name", "id"),
        filters = list(name=c("ara"),
                       strand=c("forward"),
                       posright=c("2000","40000")),
        and=TRUE,
        partialmatch = "name",
        interval= "posright")
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chávez
}
\keyword{attributes,}
\keyword{data}
\keyword{filters,}
\keyword{retrieval,}
