% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttr.R
\name{getAttr}
\alias{getAttr}
\title{Get Attributes}
\usage{
getAttr(attributes = NULL, filters = NULL, values = NULL,
  dataset = NULL, operator = "AND")
}
\arguments{
\item{attributes}{Attributes to be retrieved.}

\item{filters}{Filters to be used.}

\item{values}{Values to filter with.}

\item{dataset}{Dataset to retrieve from.}

\item{operator}{Logical operator for the filters: AND / OR.  Defaults to AND.}
}
\description{
Retrieve information based on attributes and filters from datasets.
}
\examples{
# Retrieve all genes included in all operons 
# regulated by CRP transcriptional factor.

getAttr(attributes = c("operon_name", "op_gene_names"), 
filters = c("op_tf_names"), 
values = "CRP", 
dataset = "OPERON_DM")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{attributes,}
\keyword{data}
\keyword{filters,}
\keyword{retrieval,}

