% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttr.R
\name{GetAttr}
\alias{GetAttr}
\title{Extract data from RegulonDB}
\usage{
GetAttr(attributes = NULL, filters = NULL, dataset = NULL, and = TRUE,
  interv = NULL, partialmatch = NULL)
}
\arguments{
\item{attributes}{A list or vector of attributes to be retrieved}

\item{filters}{A list of filters to be used. The name of the elements should correspond to the attribute used as
filter and the elements as the values. See details in the section of examples}

\item{dataset}{Dataset of interest. See ListDataset function to choose a dataset and ListAttributes() to see all
attributes and their descriptions available in each dataset}

\item{operator}{Logical operator for the filters: AND / OR.  Defaults to AND. See details in the section of examples}
}
\description{
This function retrieves data from RegulonDB. Data can be filtered using the filter parameter.
}
\examples{
# From "GENE" data set get all the information
Genedf<-(GetAttr(dataset="GENE"))
dim(Genedf)
head(Genedf))
# From  "GENE" data set get the posright and name
head(GetAttr(attributes=c("posright","name"),
dataset="GENE")
# From "GENE" dataset, get the gene name, strand, posright, product name and id of all genes regulated
  with name  like "ara", strand as "forward" and with a position rigth bewten 2000 and 4000

GetAttr(attributes = c("name", "strand", "posright", "product_name", "id"),
dataset = "GENE",
filters = list(name=c("ara"),
              strand=c("forward"),
              posright=c("2000","4000000")
),
and=TRUE,
partialmatch = "name",
interv="posright")
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{attributes,}
\keyword{data}
\keyword{filters,}
\keyword{retrieval,}
