% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttr.R
\name{GetAttr}
\alias{GetAttr}
\title{Extract data from RegulonDB}
\usage{
<<<<<<< HEAD
GetAttr(attributes = NULL, filters = NULL, dataset = NULL, and = TRUE)
=======
GetAttr(attributes = NULL, filters = NULL, dataset = NULL, and = TRUE,
  interv = NULL, partialmatch = NULL)
>>>>>>> 137b4a3e59e6291141d4a4eb66c64e5cdde9facc
}
\arguments{
\item{attributes}{A list or vector of attributes to be retrieved}

\item{filters}{A list of filters to be used. The name of the elements should correspond to the attribute used as
filter and the elements as the values. See details in the section of examples}

\item{dataset}{Dataset of interest. See ListDataset function to choose a dataset and ListAttributes() to see all
attributes and their descriptions available in each dataset}

\item{operator}{Logical operator for the filters: AND / OR.  Defaults to AND. See details in the section of examples}
}
\description{
This function retrieves data from RegulonDB. Data can be filtered using the filter parameter.
}
\examples{
# From "GENE" dataset, get the gene name and genome location (left position, right position and strand) of all genes regulated exclusively by Sigma Factor 19

GetAttr(attributes = c("name", "posleft", "posright", "strand"),
filters = list(sigma_factor = "Sigma19"),
dataset = "GENE")
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{attributes,}
\keyword{data}
\keyword{filters,}
\keyword{retrieval,}
