% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneRegulation.R
\name{GetGeneRegulation}
\alias{GetGeneRegulation}
\title{Get regulation}
\usage{
GetGeneRegulation(genes, format = "multirow", output.type = "TF")
}
\arguments{
\item{genes}{character vector of gene names or GIs.}

\item{format}{output format of data: "multirow", "onerow",or "table". In "multirow" format, each row represents an interaction. In "onerow", there is one row for each gene. The "table" format returns a matrix. The default output is "multirow".}

\item{output.type}{should regulators be represented as TF of GENE? The default is "TF".}
}
\description{
Given a list of genes (names, or GIs), get all transcription factors that regulate them.
}
\examples{
# Retrieve regulation of araC

GetGeneRegulation(genes = c("araC"),
format = "multirow",
)
# Retrieve regulation of several  genes, represented as a table

GetGeneRegulation(genes = c("alr", "modB","cysZ","dfp","hisM"),
format = "table",
output.type= "GENE"
)
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{tf,}
