% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetwork.R
\name{GetNetwork}
\alias{GetNetwork}
\title{Return complete regulatory network.}
\usage{
GetNetwork(regulator = NULL, type = "TF-GENE", cytograph = FALSE)
}
\arguments{
\item{regulator}{Name of TF or gene that acts as regulator. If NULL, the function retrieves all existent networks in the regulonDB.}

\item{type}{"TF-GENE", "TF-TF", "GENE-GENE"}

\item{cytograph}{If TRUE, displays network in Cytoscape. This option requires previous instalation and launch of Cytoscape.}
}
\description{
This function retrieves all the regulation networks in regulonDB between TF-TF, GENE-GENE or TF-GENE depending on the parameter 'type'
}
\examples{
# Retrieve regulation of 'araC'

GetNetwork(regulator = "AraC", type="TF-GENE")

# Retrieve all GENE-GENE networks

GetNetwork(type = "GENE-GENE")

# Retrieve TF-GENE network of AraC and display in Cytoscape

GetNetwork(regulator = "AraC", type = "TF-GENE", cytograph = TRUE)
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chávez
}
\keyword{TF,}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
