% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetwork.R
\name{GetNetwork}
\alias{GetNetwork}
\title{GetNetwork}
\usage{
GetNetwork(type = "TF-GENE")
}
\arguments{
\item{type}{tf-gene, tf-tf, gene-gene}
}
\description{
Return complete regulatory network.
}
\examples{
# Retrieve regulation of araC

GetGeneRegulation(genes = c("araC"), 
format = "multirow", 
output.type = "TF")
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{tf,}
