% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSiteSequences.R
\name{GetSiteSequences}
\alias{GetSiteSequences}
\title{Extract sequences from a set of sites.}
\usage{
GetSiteSequences(tfbs, seq.format = "fasta")
}
\arguments{
\item{tfbs}{a collection of sites previously retrieved from RegulonDB with the function GetSitesForTF()}

\item{seq.format="fasta"}{output format. Supported: "fasta", "wconsensus".}
}
\value{
a string with the sequences
}
\description{
Return the sequences for a set of binding sites.
}
\examples{
## Get the binding sites for AraC and extract their sequence in fasta format
tfbs <- GetSitesForTF(TF = "AraC")
tfbs.seq <- GetSiteSequences(tfbs) # This returns a single string with "\\n" return character
cat(tfbs.seq)
}
\author{
José & Jacques van Helden
}
