% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSitesForTF.R
\name{GetSitesForTF}
\alias{GetSitesForTF}
\title{Get the binding sites for a Transcription Factor (TF)}
\usage{
GetSitesForTF(TF, evidence = NULL)
}
\arguments{
\item{TF}{name of the transcription factor of interest}
}
\value{
A dataframe with the following columns corresponding to the TFBSs associated with the TF.
If the transcription factor does not exist, returns NA.
\itemize{
\item ID (identifier)
\item left position
\item right position
\item strand
\item sequence
}
}
\description{
Retrieve the binding sites and genome location for a given transcription factor.
}
\examples{
## Get the binding sites for AraC

tfbs <- GetSitesForTF(TF = "AraC")
}
\author{
José Alquicira Hernández, Jacques van Helden, Joselyn Chávez
}
