% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSitesForTF.R
\name{GetSitesForTF}
\alias{GetSitesForTF}
\title{Get binding sites for a selected Transcription Factot}
\usage{
GetSitesForTF(tf)
}
\arguments{
\item{tf}{name of the transcription factor of interest}
}
\value{
a dataframe with the TFBS descriptions. If the TF does not exist, return a NA.
}
\description{
Retrieve the binding sites and genome location for a given transcription factor.
}
\examples{
## Extract the binding sites for AraC
tfbs <- GetSitesForTF(tf = "AraC")
}
\author{
José Alquicira Hernández & Jacques van Helden
}
