% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSitesForTF.R
\name{GetSitesForTF}
\alias{GetSitesForTF}
\title{Get sites for a selected TF.}
\usage{
GetSitesForTF(TF)
}
\arguments{
\item{TF}{name of the transcription factor of interest}
}
\value{
a list comprising the query TF name, the number of binding sites, and a table with the TFBS descriptions.
}
\description{
Retrieve the binding sites for a given transcription factor.
}
\examples{
## Extract the binding sites for AraC
tfbs <- GetSitesForTF(TF = "AraC")
message("Got ", tfbs$nb.sites, " binding sites for ", tfbs$TF)
head(tfbs$sites) # Print the 10 first binding sistes
}
\author{
José Alquicira HErnández & Jacques van Helden
}
