% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSummary.R
\name{GetSummary}
\alias{GetSummary}
\title{Return summary of gene regulation.}
\usage{
GetSummary(geneRegulation)
}
\arguments{
\item{geneRegulation}{Result from \code{\link{GetGeneRegulation}} or vector of genes}
}
\value{
A data frame with the following columns:
\itemize{
\item The name or gene of TF
\item Regulated Genes per TF
\item Percent of regulated genes per TF
\item positive, negative or dual regulation
\item Name(s) of regulated genes
}
}
\description{
This function takes the output of \code{\link{GetGeneRegulation}} with format multirow,
onerow or table, or a vector with genes and retrieves information about the TFs and their regulated genes
}
\examples{
# Retrieve summary of  araC's regulation

araC_regulation<- GetGeneRegulation(genes = c("araC"),
                                    format = "multirow",
                                    output.type = "TF")
GetSummary(araC_regulation)

# Retrieve summary of genes 'araC' and 'modB'

GetSummary(geneRegulation = c("araC", "modB"))
GetSummary(geneRegulation = c("ECK120000050", "modB"))
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chávez
}
\keyword{networks,}
\keyword{regulation}
\keyword{retrieval,}
\keyword{summary,}
