% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets in RegulonDB database}
\usage{
list_datasets(regulondb)
}
\arguments{
\item{regulondb}{A regulondb class.}
}
\description{
This function returns a vector of all available tables from a regulondb class.
}
\examples{
download_database(getwd())
e_coli_regulondb <- regulondb(database_path=file.path(getwd(),"regulondb_sqlite3.db"), organism = "E.coli", database_version = "1", genome_version = "1")
list_datasets( e_coli_regulondb )
}
\keyword{data}
\keyword{database,}
\keyword{datasets,}
\keyword{retrieval,}
