% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAttributes.R
\name{listAttributes}
\alias{listAttributes}
\title{List attributes from dataset}
\usage{
listAttributes(dataset)
}
\arguments{
\item{dataset}{Dataset to retrieve from. Defaults to NULL.}
}
\description{
Lists attributes of datasets from RegulonDB. The result of this function may
be used as parameter values in `getAttr` function.
}
\examples{
listAttributes("TF_DM")
listAttributes("OPERON_DM")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{attributes,}
\keyword{data}
\keyword{retrieval,}
