% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListAttributes.R
\name{ListAttributes}
\alias{ListAttributes}
\title{List attributes from dataset}
\usage{
ListAttributes(dataset, comments = FALSE)
}
\arguments{
\item{dataset}{Dataset of interest.}
}
\value{
A data frame with two columns:
\itemize{
\item \code{attribute}. Name of the attribute
\item \code{description}. Description of attribute
}
}
\description{
Lists all attributes and their descriptions of a dataset from RegulonDB. The result of this function may
be used as parameter values in \code{GetAttr} function.
}
\examples{
ListAttributes("TF")
ListAttributes("OPERON")
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
\keyword{attributes,}
\keyword{data}
\keyword{retrieval,}
