% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{List available datasets in RegulonDB database}
\usage{
listDatasets()
}
\arguments{
\item{No}{parameters are used.}
}
\description{
This function returns a vector of all available datasets.
}
\examples{
listDatasets()
}
\author{
Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, Jose Alquicira Hernandez
}
\keyword{data}
\keyword{database}
\keyword{datasets,}
\keyword{retrieval,}

