% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonExistingIntervals.R
\name{NonExistingIntervals}
\alias{NonExistingIntervals}
\title{Constructs a logical condition to query database}
\usage{
NonExistingIntervals(filters, interval, operator, partialmatch)
}
\arguments{
\item{filters}{List of filters to be used. The names should correspond to the attribute and the values correspond to the condition for selection.}

\item{interval}{the filters whose values will be considered as interval}

\item{operator}{A string indicating if all the filters (AND) or some of them (OR) should be met}
}
\value{
A string with a sql logical condition to query the dataset
}
\description{
Given a list of filters, this function builds a logical condition to query database using intervals.
The output is used in \code{\link{BuildCondition}}.
}
\examples{
NonExistingIntervals(filters = list(name="ara",strand = "for"),
                     interval = NULL,
                     operator = "AND",
                     partialmatch = c("name", "strand"))
}
\author{
Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández
}
