% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dna_objects.R
\name{plot_dna_objects}
\alias{plot_dna_objects}
\title{Plot genomic elements from regulonDB}
\usage{
plot_dna_objects(
  regulondb,
  genome = "eschColi_K12",
  from = 0,
  to = 5000,
  plot_map = TRUE,
  elements = "gene"
)
}
\arguments{
\item{regulondb}{A \code{\link[=regulondb]{regulondb()}} object.}

\item{genome}{A valid UCSC genome name.}

\item{from}{A \code{integer(1)} specifying the left position.}

\item{to}{A \code{integer(1)} specifying the right position.}

\item{plot_map}{A \code{logical(1)}. By default \code{TRUE}, if \code{FALSE} returns a
\code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges-class()}} object with the elements found.}

\item{elements}{A character vector specifying which annotation elements to
plot. It can be any from: \code{"-10 promoter box"}, \code{"-35 promoter box"},
\code{"gene"}, \code{"promoter"}, \code{"Regulatory Interaction"}, \code{"sRNA interaction"},
or \code{"terminator"}.}
}
\value{
A plot with genomic elements found within a genome region,
including genes and regulators.
}
\description{
Plot genomic elements from regulonDB
}
\examples{
## Connect to the RegulonDB database if necessary
if(!exists('regulondb_conn')) regulondb_conn <- connect_database()

## Build the regulondb object
e_coli_regulondb <-
    regulondb(
        database_conn = regulondb_conn,
        organism = "chr",
        database_version = "1",
        genome_version = "1"
    )

## Plot genes from E. coli using default parameters
plot_dna_objects(e_coli_regulondb)

## Plot genes providing Genomic Ranges
plot_dna_objects(e_coli_regulondb, from = 5000, to = 10000)

## Plot aditional elements within Genomic Ranges
plot_dna_objects(e_coli_regulondb, from = 5000, to = 10000, elements = c("gene", "promoter"))
}
\author{
Joselyn Chavez
}
