% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regutools-package.R
\docType{package}
\name{regutools-package}
\alias{regutools}
\alias{regutools-package}
\title{regutools: regutools: an R package for data extraction from RegulonDB}
\description{
This package provides an interface to the RegulonDB
    http://regulondb.ccg.unam.mx/ database such that you can query it using
    R and easily work with Bioconductor packages. This package was created
    as a collaboration by members of the Community of Bioinformatics
    Software Developers (CDSB in Spanish) https://comunidadbioinfo.github.io/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ComunidadBioInfo/regutools}
  \item Report bugs at \url{https://github.com/ComunidadBioInfo/regutools/issues}
}

}
\author{
\strong{Maintainer}: Joselyn Chavez \email{joselynchavezf@gmail.com} (\href{https://orcid.org/0000-0002-4974-4591}{ORCID})

Authors:
\itemize{
  \item Carmina Barberena-Jonas \email{car.barjon@gmail.com} (\href{https://orcid.org/0000-0001-7413-638X}{ORCID})
  \item Jesus E. Sotelo-Fonseca \email{jemilianosf@gmail.com} (\href{https://orcid.org/0000-0003-1600-2396}{ORCID})
  \item Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})
  \item Alejandro Reyes \email{alejandro.reyes.ds@gmail.com} (\href{https://orcid.org/0000-0001-8717-6612}{ORCID})
}

Other contributors:
\itemize{
  \item Jose Alquicira-Hernandez \email{joseah@lcg.unam.mx} (\href{https://orcid.org/0000-0002-9049-7780}{ORCID}) [contributor]
  \item Heladia Salgado \email{heladia@ccg.unam.mx} (\href{https://orcid.org/0000-0002-3166-5801}{ORCID}) [contributor]
}

}
\keyword{internal}
