% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regutools-package.R
\docType{package}
\name{regutools-package}
\alias{regutools}
\alias{regutools-package}
\title{regutools: regutools: an R package for data extraction from RegulonDB}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

RegulonDB has collected, harmonized and centralized data 
    from hundreds of experiments for nearly two decades and is considered
    a point of reference for transcriptional regulation in Escherichia coli K12. 
    Here, we present the regutools R package to facilitate programmatic 
    access to RegulonDB data in computational biology. regutools provides 
    researchers with the possibility of writing reproducible workflows 
    with automated queries to RegulonDB. The regutools package serves as
    a bridge between RegulonDB data and the Bioconductor ecosystem by 
    reusing the data structures and statistical methods powered by other 
    Bioconductor packages. We demonstrate the integration of regutools 
    with Bioconductor by analyzing transcription factor DNA binding sites 
    and transcriptional regulatory networks from RegulonDB. We anticipate 
    that regutools will serve as a useful building block in our progress
    to further our understanding of gene regulatory networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ComunidadBioInfo/regutools}
  \item Report bugs at \url{https://support.bioconductor.org/t/regutools}
}

}
\author{
\strong{Maintainer}: Joselyn Chavez \email{joselynchavezf@gmail.com} (\href{https://orcid.org/0000-0002-4974-4591}{ORCID})

Authors:
\itemize{
  \item Carmina Barberena-Jonas \email{car.barjon@gmail.com} (\href{https://orcid.org/0000-0001-7413-638X}{ORCID})
  \item Jesus E. Sotelo-Fonseca \email{jemilianosf@gmail.com} (\href{https://orcid.org/0000-0003-1600-2396}{ORCID})
  \item Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})
  \item Alejandro Reyes \email{alejandro.reyes.ds@gmail.com} (\href{https://orcid.org/0000-0001-8717-6612}{ORCID})
}

Other contributors:
\itemize{
  \item Jose Alquicira-Hernandez \email{joseah@lcg.unam.mx} (\href{https://orcid.org/0000-0002-9049-7780}{ORCID}) [contributor]
  \item Heladia Salgado \email{heladia@ccg.unam.mx} (\href{https://orcid.org/0000-0002-3166-5801}{ORCID}) [contributor]
}

}
\keyword{internal}
