% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regutools.R
\docType{package}
\name{regutools}
\alias{regutools}
\alias{regutools-package}
\title{regutools}
\description{
regutools extracts and processes data from RegulonDB
}
\section{Maintainer}{
 Jose Alquicira and Joselyn Chavez
}

\section{See Also}{

Report bugs at https://github.com/ComunidadBioInfo/regutools.git/issues
}

\author{
Jose Alquicira Hernandez, Joselyn Chavez, Carmina Barberena Jonas, Jesus Emiliano Sotelo Fonseca
}
